requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
distlib==0.3.6
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
filelock==3.12.2
idna==3.4
mysql-connector-python==8.0.33
Pillow==10.0.0
platformdirs==3.8.1
protobuf==3.20.3
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
virtualenvwrapper-win==1.2.7